refactor: consolidate overview widgets
