--- src/components/home/index.ts
@@ -1,2 +1,0 @@
-export { default as InsightsSection } from './InsightsSection';
-export { default as InsightCard } from './InsightCard';
