+++ src/components/overview/KPIStrip.tsx
